requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pandas>=1.3.0
Pillow>=9.0.0
requests>=2.26.0
tqdm>=4.62.0
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0 